feat: update hardhat comfig to compatible with zksync
